model_dump: Add webdriver test
Summary:
The unit test previously only tested that the dump could complete
successfully.  It was not able to verify that any JS worked properly.
Now we can test the JS as long as webdriver is installed.

Tweaked the implementation of Hider a bit to make it easier for tests to
find and open them.

I disabled the tests by default since I don't want to deal with
webdriver in CI.  Enable them with the environment variable
RUN_WEBDRIVER=1.

We could make the tests use headless mode, but it's kind of fun to watch
them run.

Add a test to verify that tensor memory computation is working for the
simple model.

Test Plan:
Ran the test.

ghstack-source-id: 89889b72c3759ae9d3cc45c23d39b4f7eed54d9a
Pull Request resolved: https://github.com/pytorch/pytorch/pull/60701