[TensorExpr] Add a mechanism to register custom TS->NNC lowerings in TensorExprKernel.
The lowerings are stored as a map c10::Symbol -> std::function and the
signature of thoese functions match the signature of
`computeOperandValue`. Custom lowerings have higher priority over the
standard ones, i.e. we can redefine how a given op is lowered.

In general this feature is aimed at unblocking users whose models
contain ops that are not yet supported by NNC - it allows to quickly add
a custom lowering for a given op.

ghstack-source-id: 0e2b449368cd0d6f1129a4c2646549223b79d803
Pull Request resolved: https://github.com/pytorch/pytorch/pull/60804