Make Future store weak pointers to storages
In https://github.com/pytorch/pytorch/pull/60470 we made Future store Storages rather than store references to their DataPtrs (because these references could go stale...). However this meant that the Future could keep the Storage alive, and thus keep its memory allocated, even after the user was done with it. We fix it here by instead storing a weak ptr to that Storage (well, in fact to the StorageImpl, but it's the same).

Differential Revision: [D29454104](https://our.internmc.facebook.com/intern/diff/D29454104/)

ghstack-source-id: 132650255
Pull Request resolved: https://github.com/pytorch/pytorch/pull/60943